Remove Inspector command and code as it's sunset
